Refactor plain CLI default command helpers
