+++ packages/prime/src/prime_cli/commands/gepa.py
@@ -2,32 +2,13 @@
 
 import typer
 
-from ..utils import PlainAwareTyperGroup, PlainTyper, get_console
+from ..utils import DefaultCommandGroup, PlainTyper, get_console
 from ..verifiers_bridge import is_help_request, print_gepa_run_help, run_gepa_passthrough
 
 console = get_console()
 
 
-class DefaultGroup(PlainAwareTyperGroup):
-    def __init__(self, *args, default_cmd_name: str = "run", **kwargs):
-        super().__init__(*args, **kwargs)
-        self.default_cmd_name = default_cmd_name
-
-    def parse_args(self, ctx, args):
-        if not args:
-            return super().parse_args(ctx, args)
-
-        decision_args = [arg for arg in args if arg != "--plain"]
-        if not decision_args:
-            return super().parse_args(ctx, args)
-
-        if decision_args[0] in ("--help", "-h"):
-            return super().parse_args(ctx, args)
-        if decision_args[0] in self.commands:
-            return super().parse_args(ctx, args)
-        args = [self.default_cmd_name] + list(args)
-        return super().parse_args(ctx, args)
-
+class DefaultGroup(DefaultCommandGroup):
     def format_usage(self, ctx, formatter):
         formatter.write_usage(
             ctx.command_path,
